Move PaymentInfoBlock to own file
